+++ components/Layout/index.ts
@@ -1,0 +1,1 @@
+export { ThreeColumnLayout } from './ThreeColumnLayout';
