+++ components/ProgressBar/index.ts
@@ -1,0 +1,1 @@
+export { ProgressBar } from './ProgressBar';
