+++ components/QuizReportSidebar/index.ts
@@ -1,0 +1,1 @@
+export { QuizReportSidebar } from './QuizReportSidebar';
